[exp] svrg exp, add two situation choices, generic svrg and not generic svrg
